+++ CHANGELOG.rst
@@ -1,5 +1,12 @@
 Changelog for package marvelmind_nav
 
+1.0.9 (2020-05-28)
+------------------
+* Merge branch 'master' of https://smoker77@bitbucket.org/marvelmind_robotics/ros_marvelmind_package.git
+* Added waypoints, quality, telemetry messages. Baudrate setup via command
+  line. Relative topic names for namespaces.
+* Contributors: smoker77
+
 1.0.8 (2018-03-15)
 ------------------